Makefile: pass KEYDIR when set to sunxi-spl.bin mkimage
From: James Hilliard <james.hilliard1@gmail.com>

Provide private keys to build sunxi-spl.bin

Currently we pass this for u-boot-spl.kwb targets, however when
building sunxi-spl.bin in the TOC0 format, we may also need to
specify a KEYDIR.
As such we should also pass this when set to mkimage for the
sunxi-spl.bin target.

Tested on Whatsminer H616 board.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>
Signed-off-by: Richard Genoud <richard.genoud@bootlin.com>